Moved IAV Status to Picker Data
Accidentally added the fields to the wrong thing.
